# internal/handler/resource/host.go
package resource

import (
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"

	commodel "gin-artweb/internal/model/common"
	resomodel "gin-artweb/internal/model/resource"
	resosvc "gin-artweb/internal/service/resource"
	"gin-artweb/internal/shared/common"
	"gin-artweb/internal/shared/ctxutil"
	"gin-artweb/internal/shared/errors"
)

// HostHandler 处理主机相关的请求
// 包含日志记录和主机服务的引用
type HostHandler struct {
	log     *zap.Logger          // 日志记录器
	hostSvc *resosvc.HostService // 主机服务
}

func NewHostHandler(
	logger *zap.Logger,
	svcHost *resosvc.HostService,
) *HostHandler {
	return &HostHandler{
		log:     logger,
		hostSvc: svcHost,
	}
}

// @Summary 新增主机
// @Description 本接口用于新增主机
// @Tags 主机管理
// @Accept json
// @Produce json
// @Param request body resomodel.HostUpsertDTO true "创建主机请求"
// @Success 201 {object} resomodel.HostResp "创建主机成功"
// @Failure 400 {object} errors.Error "请求参数错误"
// @Failure 500 {object} errors.Error "服务器内部错误"
// @Router /api/v1/resource/host [post]
// @Security ApiKeyAuth
func (h *HostHandler) CreateHost(c *gin.Context) {
	startTime := time.Now()
	ctx := c.Request.Context()
	log := ctxutil.NewLogger(h.log, ctx)
	log.Info("新增主机:开始执行")

	var req resomodel.HostUpsertDTO
	if !common.ShouldBind(c, log, &req, "新增主机:绑定创建主机请求参数失败") {
		return
	}

	m, err := h.hostSvc.CreateHost(ctx, req)
	if err != nil {
		log.Error(
			"创建主机:执行失败",
			zap.Error(err),
			zap.Object("host_upsert_dto", &req),
			zap.Duration("total_duration", time.Since(startTime)),
		)
		errors.RespondWithError(c, err)
		return
	}

	log.Info(
		"创建主机:执行成功",
		zap.Uint32("host_id", m.ID),
		zap.Duration("total_duration", time.Since(startTime)),
	)

	mo := resomodel.HostModelToStandardOut(*m)
	c.JSON(http.StatusCreated, &resomodel.HostResp{
		Code: http.StatusCreated,
		Data: *mo,
	})
}

// @Summary 更新主机
// @Description 本接口用于更新指定ID的主机
// @Tags 主机管理
// @Accept json
// @Produce json
// @Param id path uint true "主机编号"
// @Param request body resomodel.HostUpsertDTO true "更新主机请求"
// @Success 200 {object} resomodel.HostResp "更新主机成功"
// @Failure 400 {object} errors.Error "请求参数错误"
// @Failure 404 {object} errors.Error "主机未找到"
// @Failure 500 {object} errors.Error "服务器内部错误"
// @Router /api/v1/resource/host/{id} [put]
// @Security ApiKeyAuth
func (h *HostHandler) UpdateHost(c *gin.Context) {
	startTime := time.Now()
	ctx := c.Request.Context()
	log := ctxutil.NewLogger(h.log, ctx)
	log.Info("更新主机:开始执行")

	var uri commodel.IDUri
	if !common.ShouldBindUri(c, log, &uri, "更新主机:绑定更新主机ID参数失败") {
		return
	}

	var req resomodel.HostUpsertDTO
	if !common.ShouldBind(c, log, &req, "更新主机:绑定更新主机请求参数失败") {
		return
	}

	m, err := h.hostSvc.UpdateHostById(ctx, uri.ID, req)
	if err != nil {
		log.Error(
			"更新主机:执行失败",
			zap.Error(err),
			zap.Uint32("host_id", uri.ID),
			zap.Object("host_upsert_dto", &req),
			zap.Duration("total_duration", time.Since(startTime)),
		)
		errors.RespondWithError(c, err)
		return
	}

	log.Info(
		"更新主机:执行成功",
		zap.Uint32("host_id", uri.ID),
		zap.Duration("total_duration", time.Since(startTime)),
	)

	mo := resomodel.HostModelToStandardOut(*m)
	c.JSON(http.StatusOK, &resomodel.HostResp{
		Code: http.StatusOK,
		Data: *mo,
	})
}

// @Summary 删除主机
// @Description 本接口用于删除指定ID的主机
// @Tags 主机管理
// @Accept json
// @Produce json
// @Param id path uint true "主机编号"
// @Success 200 {object} commodel.MapAPIResp "删除成功"
// @Failure 400 {object} errors.Error "请求参数错误"
// @Failure 404 {object} errors.Error "主机未找到"
// @Failure 500 {object} errors.Error "服务器内部错误"
// @Router /api/v1/resource/host/{id} [delete]
// @Security ApiKeyAuth
func (h *HostHandler) DeleteHost(c *gin.Context) {
	startTime := time.Now()
	ctx := c.Request.Context()
	log := ctxutil.NewLogger(h.log, ctx)
	log.Info("删除主机:开始执行")

	var uri commodel.IDUri
	if !common.ShouldBindUri(c, log, &uri, "删除主机:绑定删除主机ID参数失败") {
		return
	}

	err := h.hostSvc.DeleteHostById(ctx, uri.ID)
	if err != nil {
		log.Error(
			"删除主机:执行失败",
			zap.Error(err),
			zap.Uint32("host_id", uri.ID),
			zap.Duration("total_duration", time.Since(startTime)),
		)
		errors.RespondWithError(c, err)
		return
	}

	log.Info(
		"删除主机:执行成功",
		zap.Uint32("host_id", uri.ID),
		zap.Duration("total_duration", time.Since(startTime)),
	)

	c.JSON(commodel.NoDataResp.Code, commodel.NoDataResp)
}

// @Summary 查询主机
// @Description 本接口用于查询指定ID的主机
// @Tags 主机管理
// @Accept json
// @Produce json
// @Param id path uint true "主机编号"
// @Success 200 {object} resomodel.HostResp "获取主机详情成功"
// @Failure 400 {object} errors.Error "请求参数错误"
// @Failure 404 {object} errors.Error "主机未找到"
// @Failure 500 {object} errors.Error "服务器内部错误"
// @Router /api/v1/resource/host/{id} [get]
// @Security ApiKeyAuth
func (h *HostHandler) GetHost(c *gin.Context) {
	startTime := time.Now()
	ctx := c.Request.Context()
	log := ctxutil.NewLogger(h.log, ctx)

	var uri commodel.IDUri
	if !common.ShouldBindUri(c, log, &uri, "查询主机:绑定查询主机ID参数失败") {
		return
	}

	m, err := h.hostSvc.FindHostById(ctx, uri.ID)
	if err != nil {
		log.Error(
			"查询主机:执行失败",
			zap.Error(err),
			zap.Uint32("host_id", uri.ID),
			zap.Duration("total_duration", time.Since(startTime)),
		)
		errors.RespondWithError(c, err)
		return
	}

	mo := resomodel.HostModelToStandardOut(*m)
	c.JSON(http.StatusOK, &resomodel.HostResp{
		Code: http.StatusOK,
		Data: *mo,
	})
}

// @Summary 查询主机列表
// @Description 本接口用于查询主机列表
// @Tags 主机管理
// @Accept json
// @Produce json
// @Param request query resomodel.ListHostDTO false "查询参数"
// @Success 200 {object} resomodel.PagHostResp "成功返回主机列表"
// @Failure 400 {object} errors.Error "请求参数错误"
// @Failure 500 {object} errors.Error "服务器内部错误"
// @Router /api/v1/resource/host [get]
// @Security ApiKeyAuth
func (h *HostHandler) ListHost(c *gin.Context) {
	startTime := time.Now()
	ctx := c.Request.Context()
	log := ctxutil.NewLogger(h.log, ctx)

	var req resomodel.ListHostDTO
	if !common.ShouldBindQuery(c, log, &req, "查询主机列表:绑定查询主机列表参数失败") {
		return
	}

	page, size := req.StandardModelQuery.GetPageParam()
	total, ms, err := h.hostSvc.ListHost(ctx, page, size, req)
	if err != nil {
		log.Error(
			"查询主机列表:执行失败",
			zap.Error(err),
			zap.Int("page", page),
			zap.Int("size", size),
			zap.Object("list_host_dto", &req),
			zap.Duration("total_duration", time.Since(startTime)),
		)
		errors.RespondWithError(c, err)
		return
	}

	mbs := resomodel.ListHostModelToStandardOut(ms)
	c.JSON(http.StatusOK, &resomodel.PagHostResp{
		Code: http.StatusOK,
		Data: commodel.NewPag(page, size, total, mbs),
	})
}

func (h *HostHandler) LoadRouter(r *gin.RouterGroup) {
	r.POST("/host", h.CreateHost)
	r.PUT("/host/:id", h.UpdateHost)
	r.DELETE("/host/:id", h.DeleteHost)
	r.GET("/host/:id", h.GetHost)
	r.GET("/host", h.ListHost)
}